Move some functions to a separate core package
